Merge_PDFs_Django_Web_App
ABOUT THE TOPIC :
This Django-based web app allows users to merge multiple PDF files into one. It features a simple HTML interface with a file upload button and a "Merge PDFs" action. The merged PDF is generated on the backend using PyPDF2 and provided as a downloadable file.

YouTube Link : https://youtu.be/4PWOzObRU0A