Image_To_Text
ABOUT THE TOPIC :
This Django-based program allows users to upload an image, extract text using Tesseract OCR, and display the results instantly on a sleek web interface. The extracted text is also saved to a database for future reference. With a user-friendly design, including an image preview and a dark blue theme, this tool makes text extraction simple and efficient! 🚀

YouTube Link :  https://youtu.be/DowEoCNxw_0